requirements.txt
# Core dependencies for DeepResearcher
openai>=1.0.0
google-generativeai>=0.3.0
google-genai>=0.2.0
boto3>=1.28.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0

# For token calculation and utilities
transformers>=4.30.0
tiktoken>=0.5.0

# For HuggingFace Hub utilities (upload_to_hf_model.py, upload_to_hf_dataset.py)
huggingface_hub>=0.20.0

# Optional: For checkpoint conversion utilities (utils/convert_checkpoint.py)
# torch>=2.0.0

# For SFT training (LLaMA-Factory)
# See train/sft/LLaMA-Factory/requirements.txt for full SFT dependencies
